+++ docs/docs/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Introduced integration tests using Playwright
 - Refactor & Split up Anubis into cmd and lib.go
 - Fixed bot check to only apply if address range matches
+- Fix default difficulty setting that was broken in a refactor
 
 ## v1.14.2